+++ drivers/mfd/Kconfig
@@ -41,6 +41,7 @@ source "drivers/mfd/Kconfig.npm6001"
 source "drivers/mfd/Kconfig.pca9422"
 source "drivers/mfd/Kconfig.pf1550"
 source "drivers/mfd/Kconfig.rv3032_mfd"
+source "drivers/mfd/Kconfig.sc18is606_mfd"
 source "drivers/mfd/Kconfig.tle9104"
 # zephyr-keep-sorted-stop
